+++ src/apps/settings/modules/WindowManager/border/infra.tsx
@@ -2,6 +2,7 @@ import { SettingsOption, SettingsSubGroup } from '../../../components/SettingsBo
 import { InputNumber, Switch } from 'antd';
 
 import { useAppDispatch, useAppSelector, useDispatchCallback } from '../../shared/utils/infra';
+
 import { BorderSelectors } from '../../shared/store/app/selectors';
 import { BorderActions } from './app';