+++ src/background/seelen_wm/handler.rs
@@ -1,9 +1,9 @@
 use serde::Deserialize;
 use windows::Win32::{
-    Foundation::HWND,
+    Foundation::{HWND, RECT},
     UI::WindowsAndMessaging::{
-        SetWindowPos, SWP_ASYNCWINDOWPOS, SWP_NOACTIVATE, SWP_NOCOPYBITS, SWP_NOOWNERZORDER,
-        SWP_NOSENDCHANGING, SWP_NOZORDER,
+        SWP_ASYNCWINDOWPOS, SWP_NOACTIVATE, SWP_NOCOPYBITS, SWP_NOOWNERZORDER, SWP_NOSENDCHANGING,
+        SWP_NOZORDER,
     },
 };
 
@@ -31,24 +31,22 @@ pub fn set_window_position(hwnd: isize, rect: Rect) -> Result<(), String> {
 
     WindowsApi::unmaximize_window(hwnd);
     let shadow = WindowsApi::shadow_rect(hwnd)?;
-    let result = unsafe {
-        SetWindowPos(
-            hwnd,
-            HWND(0),
-            rect.left + shadow.left,
-            rect.top + shadow.top,
-            rect.right + shadow.right + shadow.left.abs(),
-            rect.bottom + shadow.bottom + shadow.top.abs(),
-            SWP_NOACTIVATE
-                | SWP_NOCOPYBITS
-                | SWP_NOZORDER
-                | SWP_NOOWNERZORDER
-                | SWP_ASYNCWINDOWPOS
-                | SWP_NOSENDCHANGING,
-        )
-    };
-
-    log_if_error(result);
+    WindowsApi::set_position(
+        hwnd,
+        None,
+        &RECT {
+            top: rect.top + shadow.top,
+            left: rect.left + shadow.left,
+            right: rect.right + shadow.right,
+            bottom: rect.bottom + shadow.bottom,
+        },
+        SWP_NOACTIVATE
+            | SWP_NOCOPYBITS
+            | SWP_NOZORDER
+            | SWP_NOOWNERZORDER
+            | SWP_ASYNCWINDOWPOS
+            | SWP_NOSENDCHANGING,
+    )?;
     Ok(())
 }